requirements.txt
Django==5.2
Pillow==10.2.0
mysqlclient==2.2.0
django-crispy-forms==2.0
django-filter==23.2
reportlab==4.0.4
gunicorn==21.2.0
whitenoise==6.5.0
python-dateutil==2.8.2
pytz==2023.3

cryptography==42.0.5

djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
